requirements.txt
pytest
redis
PyYaml